Move SqlServer docker helper to TestUtils. Also move TaskExtensions.
