+++ src/SqlStreamStore.MsSql.V3.Tests/MsSqlStreamStoreV3Fixture.cs
@@ -1,11 +1,8 @@
 namespace SqlStreamStore
 {
     using System;
-    using System.Collections.Generic;
     using System.Data.SqlClient;
-    using System.Threading;
     using System.Threading.Tasks;
-    using SqlStreamStore.Infrastructure;
 
     public class MsSqlStreamStoreV3Fixture : StreamStoreAcceptanceTestFixture
     {
@@ -15,7 +12,7 @@ namespace SqlStreamStore
         private readonly string _databaseNameOverride;
         private readonly bool _deleteDatabaseOnDispose;
         private readonly string _databaseName;
-        private readonly DockerSqlServerDatabase _databaseInstance;
+        private readonly DockerMsSqlServerDatabase _databaseInstance;
 
         public MsSqlStreamStoreV3Fixture(
             string schema,
@@ -28,7 +25,7 @@ namespace SqlStreamStore
             _databaseNameOverride = databaseNameOverride;
             _deleteDatabaseOnDispose = deleteDatabaseOnDispose;
             _databaseName = databaseNameOverride ?? $"sss-v3-{Guid.NewGuid():n}";
-            _databaseInstance = new DockerSqlServerDatabase(_databaseName);
+            _databaseInstance = new DockerMsSqlServerDatabase(_databaseName);
 
             ConnectionString = CreateConnectionString();
         }
@@ -123,78 +120,6 @@ namespace SqlStreamStore
             connectionStringBuilder.InitialCatalog = _databaseName;
 
             return connectionStringBuilder.ToString();
-        }
-
-        private class DockerSqlServerDatabase
-        {
-            private readonly string _databaseName;
-            private readonly DockerContainer _sqlServerContainer;
-            private const string Password = "!Passw0rd";
-            private const string Image = "microsoft/mssql-server-linux";
-            private const string Tag = "2017-CU9";
-            private const int Port = 11433;
-
-            public DockerSqlServerDatabase(string databaseName)
-            {
-                _databaseName = databaseName;
-
-                var ports = new Dictionary<int, int>
-                {
-                    { 1433, Port }
-                };
-
-                _sqlServerContainer = new DockerContainer(
-                    Image,
-                    Tag,
-                    HealthCheck,
-                    ports)
-                {
-                    ContainerName = "sql-stream-store-tests-mssql",
-                    Env = new[] { "ACCEPT_EULA=Y", $"SA_PASSWORD={Password}" }
-                };
-            }
-
-            public SqlConnection CreateConnection()
-                => new SqlConnection(CreateConnectionStringBuilder().ConnectionString);
-
-            public SqlConnectionStringBuilder CreateConnectionStringBuilder()
-                => new SqlConnectionStringBuilder($"server=localhost,{Port};User Id=sa;Password={Password};Initial Catalog=master");
-
-            public async Task CreateDatabase(CancellationToken cancellationToken = default)
-            {
-                await _sqlServerContainer.TryStart(cancellationToken).WithTimeout(3 * 60 * 1000);
-
-                using(var connection = CreateConnection())
-                {
-                    await connection.OpenAsync(cancellationToken).NotOnCapturedContext();
-
-                    var createCommand = $@"CREATE DATABASE [{_databaseName}]
-ALTER DATABASE [{_databaseName}] SET SINGLE_USER
-ALTER DATABASE [{_databaseName}] SET COMPATIBILITY_LEVEL=110
-ALTER DATABASE [{_databaseName}] SET MULTI_USER";
-
-                    using (var command = new SqlCommand(createCommand, connection))
-                    {
-                        await command.ExecuteNonQueryAsync(cancellationToken).NotOnCapturedContext();
-                    }
-                }
-            }
-
-            private async Task<bool> HealthCheck(CancellationToken cancellationToken)
-            {
-                try
-                {
-                    using(var connection = CreateConnection())
-                    {
-                        await connection.OpenAsync(cancellationToken).NotOnCapturedContext();
-
-                        return true;
-                    }
-                }
-                catch (Exception) { }
-
-                return false;
-            }
         }
     }
 }
