+++ src/SqlStreamStore.TestUtils/SqlStreamStore.TestUtils.csproj
@@ -1,7 +1,7 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <RootNamespace>SqlStreamStore</RootNamespace>
   </PropertyGroup>
 
@@ -9,6 +9,7 @@
     <PackageReference Include="Docker.DotNet" Version="3.125.2" />
     <PackageReference Include="Serilog" Version="2.7.1" />
     <PackageReference Include="Serilog.Sinks.Observable" Version="2.0.2" />
+    <PackageReference Include="System.Data.SqlClient" Version="4.6.0" />
     <PackageReference Include="System.Reactive" Version="4.1.0" />
     <PackageReference Include="xunit.core" Version="2.4.0" />
   </ItemGroup>
